requirements.txt
torch==2.7.0
torchvision==0.22.0
tensorboard==2.19.0
# torchaudio --index-url https://download.pytorch.org/whl/cu121
pillow==10.3.0
fastapi==0.110.2
uvicorn[standard]==0.29.0
numpy==2.2.5
opencv-python-headless==4.11.0.86
python-multipart==0.0.20
tqdm==4.66.4
albumentations==2.0.6
scikit-image==0.25.2